+++ package.json
@@ -43,22 +43,29 @@
     "url": "https://github.com/apparatus/fuge-runner"
   },
   "scripts": {
-    "lint": "spacey-standard",
-    "test": "npm run lint && tape test/*Test.js | tap-spec",
-    "coverage": "istanbul cover tape **/*Test.js && open ./coverage/lcov-report/index.html",
-    "coverage-check": "istanbul cover tape test/*Test.js && istanbul check-coverage",
-    "commit-check": "npm run test"
+    "lint": "echo 'linting...' && spacey-standard | snazzy",
+    "test": "tap 'test/**/*.test.js' --branches=90 --statements=90 --functions=90 --lines=90",
+    "check": "npm run test && npm run lint",
+    "coverage": "npm run test -- --cov --coverage-report=html",
+    "coveralls": "npm run test -- --cov --coverage-report=text-lcov | coveralls"
+  },
+  "nyc": {
+    "exclude": [
+      "**/test/**/*.js"
+    ]
   },
   "pre-commit": [
     "commit-check"
   ],
   "devDependencies": {
-    "istanbul": "^0.4.2",
+    "coveralls": "^2.11.14",
+    "json-stringify-safe": "^5.0.1",
+    "pre-commit": "^1.1.3",
+    "snazzy": "^5.0.0",
     "spacey-standard": "^3.0.0",
-    "pre-commit": "^1.1.1",
+    "tap": "^7.1.2",
     "proxyquire": "^1.7.10",
     "request": "^2.69.0",
-    "tap-spec": "^4.1.1",
-    "tape": "^4.2.0"
+    "tap-spec": "^4.1.1"
   }
 }
